add a separate test suite for custom functionality, push version to 0.6.0
