+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@yoolabs/class-transformer",
-  "version": "0.5.1",
+  "version": "0.6.0",
   "description": "Fork of the class-transformer package. Proper decorator-based transformation / serialization / deserialization of plain javascript objects to class constructors",
   "author": "TypeStack contributors, Jannik Mewes (YOOniversity)",
   "license": "MIT",
